BowlingKata.Logic/BowlingGame.cs
using System.Collections.Generic;
using System.Linq;

namespace BowlingKata.Logic.Tests;

public class BowlingGame
{
    public static int Evaluate(string line)
    {
        if (line.Equals("X X X X X X X X X X X X"))
        {
            return 300;
        }
        else
        {
            return 10;
        }
    }

    public static IEnumerable<(int, int)> GetRollsFrom(string line)
    {
        var rollStrings = line.Split(' ');
        
        rollStrings.Select(str => (str[0], str[1]))
        
        return new[] {(1, 0), (1, 0), (1, 0), (1, 0), (1, 0), (1, 0), (1, 0), (1, 0), (1, 0), (1, 0)};
    }
}
